use Byte instead of Int

+++ src/Brainfuck/AST.cs
@@ -12,23 +12,23 @@ public abstract record AST {
   static readonly Stream In  = Console.OpenStandardInput();
   static readonly Stream Out = Console.OpenStandardOutput();
 
-  static IO<Unit> WriteByte(int b) => IO.lift(() => { Out.WriteByte((byte) b); Out.Flush(); });
+  static IO<Unit> WriteByte(byte b) => IO.lift(() => { Out.WriteByte(b); Out.Flush(); });
   static IO<int>  ReadByte()       => IO.lift(() => In.ReadByte());
 
   public static StateT<Mem, IO, Unit> Step(AST ast) {
     return ast switch {
-      Incr => StateT.modify<IO, Mem>(m => m with { Cur = m.Cur + 1}),
-      Decr => StateT.modify<IO, Mem>(m => m with { Cur = m.Cur - 1}),
+      Incr => StateT.modify<IO, Mem>(m => m with { Cur = (byte)(m.Cur + 1)}),
+      Decr => StateT.modify<IO, Mem>(m => m with { Cur = (byte)(m.Cur - 1)}),
       Left => StateT.modify<IO, Mem>(m => m.MoveLeft()),
       Right => StateT.modify<IO, Mem>(m => m.MoveRight()),
-      Write => StateT.gets<IO, Mem, int>(m => m.Cur)
-        .Bind((int b) => StateT.liftIO<Mem, IO, Unit>(WriteByte(b))),
+      Write => StateT.gets<IO, Mem, byte>(m => m.Cur)
+        .Bind((byte b) => StateT.liftIO<Mem, IO, Unit>(WriteByte(b))),
       Read => StateT.liftIO<Mem, IO, int>(ReadByte())
-        .Bind(b => new Modify<Mem>(m => m with {Cur = b})),
+        .Bind(b => new Modify<Mem>(m => m with {Cur = (byte) b})),
       Loop(var body) => 
-        from cur in StateT.gets<IO, Mem, int>(m => m.Cur)
+        from cur in StateT.gets<IO, Mem, byte>(m => m.Cur)
         from _1 in when(cur != 0, from _2 in Eval(body)
-                                  from curLoopEnd in StateT.gets<IO, Mem, int>(m => m.Cur)
+                                  from curLoopEnd in StateT.gets<IO, Mem, byte>(m => m.Cur)
                                   from _3 in when(curLoopEnd != 0, Step(ast))
                                   select unit)
         select unit,
